--- app/src/config/MultiHttpSecurityConfig.java
@@ -1,8 +1,0 @@
-package config;
-
-//@EnableWebSecurity
-public class MultiHttpSecurityConfig {
-//    public static class ApiWebSecurityConfigurationAdapter extends WebSecurityConfigurerAdapter{
-//
-//    }
-}
