+++ app/src/resources/DataBase.java
@@ -8,39 +8,35 @@ import java.util.List;
 import java.util.logging.Logger;
 
 public class DataBase {
-    private static final Logger logger = Logger.getLogger("DataBase");
-    private static DataBase dataBaseInstance;
+    private static final Logger logger = Logger.getLogger(DataBase.class.getName());
+    private static final DataBase dataBaseInstance = new DataBase();
     private final List<InstaAccount> instaAccountArrayList = new ArrayList<>();
 
-    private final InstaAccount[] shops = new InstaAccount[]{new InstaAccount("StyleNanda shop"),
-            new InstaAccount("Zara"),
-            new InstaAccount("Forever_21"),
-            new InstaAccount("Pazzolini"),
-            new InstaAccount("awesome_shoes"),
-            new InstaAccount("Lacoste")};
-
-    public DataBase() {
+    private DataBase() {
         logger.info("filling Database...");
-        fillInstaAccountArrayList();
+        instaAccountArrayList.addAll(Arrays.asList(
+                new InstaAccount("StyleNanda shop"),
+                new InstaAccount("Zara"),
+                new InstaAccount("Forever_21"),
+                new InstaAccount("Pazzolini"),
+                new InstaAccount("awesome_shoes"),
+                new InstaAccount("Lacoste"))
+        );
     }
 
     public static DataBase getDataBaseInstance() {
-        if (dataBaseInstance == null) {
-            dataBaseInstance = new DataBase() {
-            };
-        }
         return dataBaseInstance;
     }
 
-    private void fillInstaAccountArrayList() {
-        instaAccountArrayList.addAll(Arrays.asList(shops));
+    public static void start() {
     }
 
     public String searchByInstaAccountName(String instaAccountName) {
 
         for (InstaAccount account : instaAccountArrayList) {
-            if (account.getInstaAccountName().equals(instaAccountName))
+            if (account.getInstaAccountName().equals(instaAccountName)) {
                 return account.getInstaAccountName();
+            }
         }
         return null;
     }
