+++ app/src/server/Server.java
@@ -1,6 +1,5 @@
 package server;
 
-import com.sun.net.httpserver.HttpContext;
 import com.sun.net.httpserver.HttpServer;
 
 import java.io.IOException;
@@ -12,39 +11,25 @@ import java.util.logging.Logger;
 
 public class Server {
 
-    private static final Logger logger = Logger.getLogger("Server");
-
-    private static final String defaultPage = "/";
-    private static final String instaAccountsPage = "/accounts";
-    private static final String accountSearchPage = "/accounts/search?name=awesome_shoes";
+    private static final Logger logger = Logger.getLogger(Server.class.getName());
 
     private Server() {
     }
 
     public static void start() throws IOException {
+        HttpServer httpServerInstance = HttpServer.create(new InetSocketAddress(8080), 0); // 0 - для обратной регистрации, если 0 - не ставим в очередь никаких запросов
+        ServerHandlerToSearch serverHandlerToSearch = new ServerHandlerToSearch();
+        ServerHandlerToShowAccounts serverHandlerToShowAccounts = new ServerHandlerToShowAccounts();
+        logger.info("ServerHandler created");
 
-        try {
-            HttpServer httpServerInstance = HttpServer.create(new InetSocketAddress(8080), 0); // 0 - для обратной регистрации, если 0 - не ставим в очередь никаких запросов
-            ServerHandler serverHandler = new ServerHandler();
-            logger.info("ServerHandler created");
+        ThreadPoolExecutor threadPoolExecutor = (ThreadPoolExecutor) Executors.newFixedThreadPool(10);
+        httpServerInstance.setExecutor(threadPoolExecutor);
+        logger.info("Executor created");
 
-            ThreadPoolExecutor threadPoolExecutor = (ThreadPoolExecutor) Executors.newFixedThreadPool(10);
-            httpServerInstance.setExecutor(threadPoolExecutor);
-            logger.info("Executor created");
+        httpServerInstance.createContext("/accounts", serverHandlerToShowAccounts);
+        httpServerInstance.createContext("/accounts/search?name=awesome_shoes", serverHandlerToSearch);
 
-            HttpContext defaultContext = httpServerInstance.createContext(defaultPage, serverHandler);
-            httpServerInstance.createContext(instaAccountsPage, serverHandler);
-            httpServerInstance.createContext(accountSearchPage, serverHandler);
-
-            httpServerInstance.start();
-        } catch (IOException ex) {
-            logger.warning("Unable to start server");
-            logger.warning(ex.getMessage());
-            ex.printStackTrace();
-            throw new RuntimeException(ex);
-        }
-
+        httpServerInstance.start();
         logger.info("Server started");
-
     }
 }
