+++ app/src/server/ServerHandlerToShowAccounts.java
@@ -1,0 +1,45 @@
+package server;
+
+import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
+
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.logging.Logger;
+
+
+public class ServerHandlerToShowAccounts extends ServerResponseHandler implements HttpHandler {
+
+    private static final Logger logger = Logger.getLogger(HttpHandler.class.getName());
+
+    public ServerHandlerToShowAccounts() {
+    }
+
+    @Override
+    public void handle(HttpExchange httpExchange) throws IOException {
+        String requestParamValue = null;
+
+        try {
+            if ("GET".equals(httpExchange.getRequestMethod())) {
+                requestParamValue = handleGetRequest();
+                logger.info(requestParamValue);
+            } else if ("POST".equals(httpExchange)) {
+                requestParamValue = handlePostRequest(httpExchange);
+            } else
+                httpExchange.sendResponseHeaders(405, -1);
+            handleResponse(httpExchange, requestParamValue, null);
+        } catch (IOException ex) {
+            throw new IOException();
+        }
+    }
+
+    private String handlePostRequest(HttpExchange httpExchange) {
+        String messageToPost = null;
+        logger.info("method post");
+        return messageToPost;
+    }
+
+    private String handleGetRequest() throws IOException {
+        return "Requested Instagram Accounts not found. For now you can look for 'awesome_shoes'";
+    }
+}
