+++ app/src/store/ProductItem.java
@@ -3,13 +3,10 @@ package store;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-
 public class ProductItem {
     private UUID id;
     private String title;
-    private String description;
     private BigDecimal price;
-    private int size;
 
     public ProductItem(String title, BigDecimal price) {
         this.id = UUID.randomUUID();
@@ -25,15 +22,7 @@ public class ProductItem {
         return title;
     }
 
-    public String getDescription() {
-        return description;
-    }
-
     public BigDecimal getPrice() {
         return price;
-    }
-
-    public int getSize() {
-        return size;
     }
 }
